requirements.txt
foolbox
matplotlib
numpy
pandas
PyYAML
scikit-learn
torch
torchvision
tqdm
wandb